requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
autoflake==2.3.1
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.2.0
build==1.2.2.post1
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
drf-yasg==1.21.8
executing==2.1.0
factory_boy==3.3.1
Faker==33.3.0
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.5
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipython==8.12.3
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nodeenv==1.9.1
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip-tools==7.4.1
pipdeptree==2.24.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pydantic==2.10.5
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
ruff==0.8.6
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
starkbank-ecdsa==2.2.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.1
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.18.3
